GC2D/GCConsole2 processDisappearBalloon: 98.2% -> 99.4% (restore signed alpha conversion)

+++ src/GC2D/GCConsole2.cpp
@@ -3987,8 +3987,8 @@ bool TGCConsole2::processDisappearBalloon()
 		                    unk3BC.y2);
 		unk3B0->mBounds = nextBounds;
 	} else {
-		u16 alpha = unk3B0->mAlpha - 0x30;
-		if ((s16)alpha < 0) {
+		s32 alpha = (s16)(unk3B0->mAlpha - 0x30);
+		if (alpha < 0) {
 			alpha      = 0;
 			isFinished = true;
 		}
